// src/modules/checkout/usecases/place-order/place-order.dto.ts
export interface PlaceOrderInputDto {
    clientId: string;
    products: {
        productId: string;
        quantity: number;
    }[];
}


export interface PlaceOrderOutputDto {
    id: string;
    invoiceId: string;
    status: string;
    total: number;
    products: {
        productId: string;
    }[];
}